remove inheritance
