--- Assets/VRM10/Editor/Components/Expression/ExpressionEditorBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6a64fbc26ee3c274fbfd9bce32f0b5dd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
